extract/transformをサイトごとのサブパッケージにまとめる
- pipelines/suumo/, pipelines/mansionreview/ をサブパッケージ化し、
  extract.py, transform.py, pipeline.py, __init__.py に分割
- extract/utils.py と transform/utils.py はサイト非依存の共通処理
  としてpipelines/core/utils.pyに統合
- extract/suumo.py, extract/mansionreview.py, transform/以下の
  site別モジュールは削除(extract/mansionreview_history.pyは
  Pipelineの型に乗らない独立スクリプトのため維持)
