+++ pipelines/core/utils.py
@@ -1,9 +1,13 @@
-#utilsじゃなくてcommon.pyを別で作ってもいいかも
+import multiprocessing
+from multiprocessing import Pool
+
 import numpy as np
+import pandas as pd
+import urllib
 from retry import retry
 import requests
-import urllib
-import pandas as pd
+from tqdm import tqdm
+
 
 def sigmoid(x, gain=1, offset_x=0):
     return ((np.tanh(((x+offset_x)*gain)/2)+1)/2)
@@ -23,8 +27,22 @@ def search_address(address):
 
 def get_diff_records(today,yesterday):
     diff=pd.merge(today,yesterday, on=["建物名","面積"], how ="outer", indicator=True).query(f'_merge != "both"')
-    
+
     apeared=diff[diff["_merge"]=="left_only"]
     banished=diff[diff["_merge"]=="right_only"]
 
     return apeared.merge(today,on=["建物名","面積"],how="left"),banished.merge(yesterday,on=["建物名","面積"],how="left")
+
+def get_lat_lon(addresses):
+
+    LEN_ADDRESSES=len(addresses)
+    lons=[]
+    lats=[]
+    with Pool(processes=multiprocessing.cpu_count()) as pool:
+        with tqdm(total=LEN_ADDRESSES) as pbar:
+            for result in pool.imap(search_address, addresses):
+                lons.append(result[0])
+                lats.append(result[1])
+                pbar.update(1)
+
+    return lons,lats
