+++ pipelines/mansionreview/__init__.py
@@ -1,0 +1,1 @@
+from .pipeline import MansionReviewPipeline
