+++ pipelines/suumo/__init__.py
@@ -1,0 +1,1 @@
+from .pipeline import SuumoPipeline
